move TestValidateAnndataRawCounts to test_validate.py
