+++ cellxgene_schema_cli/tests/test_validate.py
@@ -24,7 +24,7 @@ from cellxgene_schema.validate import (
     validate,
 )
 from cellxgene_schema.write_labels import AnnDataLabelAppender
-from dask.array import from_array
+from dask_expr.array import from_array
 from fixtures.examples_validate import adata as adata_valid
 from fixtures.examples_validate import (
     adata_minimal,
@@ -39,7 +39,7 @@ from fixtures.examples_validate import (
     h5ad_valid,
     visium_library_id,
 )
-from numpy import ndarray
+from numpy import ndarray, zeros
 from scipy import sparse
 from scipy.sparse import spmatrix
 
@@ -1336,3 +1336,52 @@ class TestIsRaw:
         validator = self.create_validator(data, "unknown")
         assert validator._has_valid_raw() is False
         assert validator.errors == [f"Unknown encoding for matrix X. {ERROR_SUFFIX_SPARSE_FORMAT}"]
+
+
+class TestValidateAnndataRawCounts:
+    """
+    Paired Assays (measures both accessibility and gene expression) require raw counts to be present and validated.
+    Unpaired Assays (measures only accessibility) do not require raw counts to be present.
+
+    This is the purview of the AnnData Validator, not the AtacValidator, which is primarily focused on the fragment file.
+    """
+
+    def test_paired_requires_raw_validation(self, atac_anndata, tmpdir):
+        # 10x multiome (EFO:0030059) is paired (both ATAC and RNA single cell sequencing)
+        atac_anndata.obs["assay_ontology_term_id"] = ["EFO:0030059"] * 3
+
+        # use a valid count matrix (as dask array)
+        X = atac_anndata.X
+
+        # validate with AnnData Validator
+        validator = Validator(ignore_labels=True)
+        validator._set_schema_def()
+
+        # do validation with a valid count matrix
+        atac_anndata.X = from_array(X.astype("float32"))
+        validator.adata = atac_anndata
+        validator.reset()
+        validator._validate_raw()
+        assert validator.errors == []
+
+        # do validation with an invalid count matrix
+        atac_anndata.X = from_array(zeros(X.shape).astype("float32"))
+        validator.adata = atac_anndata
+        validator.reset()
+        validator._validate_raw()
+        assert len(validator.errors) > 0
+
+    def test_unpaired_skips_raw_validation(self, atac_anndata, tmpdir):
+        # scATAC-seq (EFO:0010891) is unpaired paired
+        atac_anndata.obs["assay_ontology_term_id"] = ["EFO:0010891"] * 3
+
+        # remove matrix - it shouldn't be required
+        del atac_anndata.X
+
+        # check that validation passes even without matrix
+        validator = Validator(ignore_labels=True)
+        validator._set_schema_def()
+        validator.adata = atac_anndata
+        validator.reset()
+        validator._validate_raw()
+        assert validator.errors == []
